src/presenters/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.7
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.0.4
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-enum==1.5.1
pdfkit==0.6.1
PyJWT==1.7.1
python-dotenv==0.10.5
pytz==2019.3
PyYAML==5.4
six==1.14.0
Werkzeug==0.16.0
